+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [1.41.0] - 2024-27-08
+
+### Added
+
+- Added support for handling the below operators in targeting and whitelisting i.e.
+
+  - Greater than
+  - Greater than equal to
+  - Less than
+  - Less than equal to
 
 ## [1.40.0] - 2024-04-08